hal_g6: halbb: avoid mask == 0 addr == 0
UBSAN reports errors due to shifts of 32 bits on u32s

The cause of this is mask == 0 passed to halbb_cal_bit_shift().
My observations show that mask == 0 is coming only with addr == 0,
suggesting there is common code with "optional" registers that are not
being filled in on some models/variants.

So this minimal change only avoids that single scenario, assuming the
rest of the code knows what it's doing.

Signed-off-by: Greg Whiteley <greg.whiteley@gmail.com>